chronos-gateway/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyyaml>=6.0
aiohttp>=3.9.0
pydantic>=2.0.0
psutil>=5.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Windows specific
pywin32>=306; sys_platform == 'win32'

# Optional
jinja2>=3.1.0
